apps/auto_scaler/requirements.txt
boto3>=1.17.9
clearml>=1.2.0rc0
google-api-python-client~=2.15
plotly>=4.14.3
six>=1.15.0
pandas~=1.3  # Need for plotly express
